Data_scraping /requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
google-cloud-bigquery>=3.25.0
